requirements.txt
archiveis==0.0.9
archivenow==2020.7.18.12.19.44
bleach==3.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.3
coveralls==2.2.0
cycler==0.10.0
docopt==0.6.2
docutils==0.16
flake8==3.8.4
Flask==1.1.2
idna==2.10
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==21.5.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.4
packaging==20.8
pandas==1.1.4
Pillow==8.0.1
pkginfo==1.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
readme-renderer==28.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
tqdm==4.54.1
twine==3.2.0
typing-extensions==3.7.4.3
urllib3==1.26.2
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.4.0
